fix header overlap with navbar in house shcedule
