+++ client/src/api/axiosInstance.js
@@ -1,33 +1,67 @@
-// src/api/axiosInstance.js
-import axios from 'axios';
+import { env } from "@/config/env";
+import axios from "axios";
+
+
 
 const axiosInstance = axios.create({
-  baseURL: 'http://it31-starterpack.sit.kmutt.ac.th/samosit/it31starterpack',
+  // baseURL: env.API_SERVER_URL,
+  baseURL: "http://it31-starterpack.sit.kmutt.ac.th/samosit/it31starterpack",
   timeout: 10000,
-  headers: {
-    'Content-Type': 'application/json',
-  },
+  withCredentials: true,
 });
 
-// Request interceptor
-axiosInstance.interceptors.request.use(
-  (config) => {
-    console.log('Request:', config.url);
-    return config;
-  },
-  (error) => Promise.reject(error)
-);
+// ตัวแปรควบคุมว่า request refresh กำลังทำงานอยู่หรือไม่
+let isRefreshing = false;
+let failedQueue = [];
+
+const processQueue = (error, token = null) => {
+  failedQueue.forEach(prom => {
+    if (error) {
+      prom.reject(error);
+    } else {
+      prom.resolve(token);
+    }
+  });
+
+  failedQueue = [];
+};
 
 // Response interceptor
 axiosInstance.interceptors.response.use(
-  (response) => {
-    console.log('Response:', response.data);
-    return response;
-  },
-  (error) => {
-    console.error('API Error:', error.response?.data || error.message);
+  response => response,
+  async error => {
+    const originalRequest = error.config;
+
+    if (error.response?.status === 401 && !originalRequest._retry) {
+      if (isRefreshing) {
+        return new Promise((resolve, reject) => {
+          failedQueue.push({ resolve, reject });
+        })
+          .then(() => axiosInstance(originalRequest))
+          .catch(err => Promise.reject(err));
+      }
+
+      originalRequest._retry = true;
+      isRefreshing = true;
+
+      try {
+        await axiosInstance.post("/users/refresh");
+
+        processQueue(null);
+        return axiosInstance(originalRequest);
+      } catch (refreshError) {
+        processQueue(refreshError, null);
+
+        window.location.href = "/login"; 
+        return Promise.reject(refreshError);
+      } finally {
+        isRefreshing = false;
+      }
+    }
+
     return Promise.reject(error);
   }
 );
+
 
 export default axiosInstance;
