--- client/src/api/axiosInstant.js
@@ -1,65 +1,0 @@
-import { env } from "@/config/env";
-import axios from "axios";
-
-
-const axiosInstance = axios.create({
-  baseURL: env.API_SERVER_URL,
-  timeout: 10000,
-  withCredentials: true,
-});
-
-// ตัวแปรควบคุมว่า request refresh กำลังทำงานอยู่หรือไม่
-let isRefreshing = false;
-let failedQueue = [];
-
-const processQueue = (error, token = null) => {
-  failedQueue.forEach(prom => {
-    if (error) {
-      prom.reject(error);
-    } else {
-      prom.resolve(token);
-    }
-  });
-
-  failedQueue = [];
-};
-
-// Response interceptor
-axiosInstance.interceptors.response.use(
-  response => response,
-  async error => {
-    const originalRequest = error.config;
-
-    if (error.response?.status === 401 && !originalRequest._retry) {
-      if (isRefreshing) {
-        return new Promise((resolve, reject) => {
-          failedQueue.push({ resolve, reject });
-        })
-          .then(() => axiosInstance(originalRequest))
-          .catch(err => Promise.reject(err));
-      }
-
-      originalRequest._retry = true;
-      isRefreshing = true;
-
-      try {
-        await axiosInstance.post("/users/refresh");
-
-        processQueue(null);
-        return axiosInstance(originalRequest);
-      } catch (refreshError) {
-        processQueue(refreshError, null);
-
-        window.location.href = "/login"; 
-        return Promise.reject(refreshError);
-      } finally {
-        isRefreshing = false;
-      }
-    }
-
-    return Promise.reject(error);
-  }
-);
-
-
-export default axiosInstance;
